Moved dialogs to windows folder and uis to ui folder
