+++ src/windows/LedIndicatorWidget.py
