+++ src/windows/powermeter_dialog.py
