+++ src/windows/settings_dialog.py
