+++ src/windows/translator_controller.py
