+++ src/windows/ui/powermeter_dialog.ui
