+++ src/windows/ui/settings_dialog.ui
