+++ src/windows/ui/warn_dialog.ui
